refactor vs specific logic out of kits.ts
